v0.4.7
v0.4.7 - added a utility function for robust S3 retrieval (retries with exponential backoff) and implemented it within the core inspection functions in an attempt to resolve the ongoing S3 curl request failures. Also allowed importance threshold to be specified as a string rather than an instance of the enum class. Also added negative job slicing to all n_jobs arguments.