v0.4.8
v0.4.8 - Adds a check for a new data object available in NWB 2.5 called 'Images'. Specifically, for the field named "order_of_images", which specifies the order of the images within the Images object, we check that the "order_of_images" contains references to all of the images only once that are in the Images object
"order_of_images" contains no other images (e.g., from another object).